doing it the celery way
