+++ celery/app.py
@@ -1,0 +1,5 @@
+from celery import Celery
+
+celery = Celery("ml_app",
+                broker="redis://localhost:6379/0",
+                backend="redis://localhost:6379/1")
