+++ celery/requirements.txt
@@ -1,0 +1,12 @@
+torch
+flask
+pillow
+mysql-connector
+opencv-python-headless
+torchvision
+colorspacious
+colormath
+gdown
+gradio
+matplotlib
+joblib
