--- ml_app/.dockerignore
@@ -1,1 +1,0 @@
-dataset/
