--- ml_app/models/encoder/palette.py
@@ -1,41 +1,0 @@
-palette_rgb = [
-    (255, 209, 223),
-    (210, 66, 71),
-    (241, 89, 42),
-    (245, 239, 221),
-    (251, 214, 78),
-    (0, 121, 71),
-    (173, 216, 230),
-    (26, 72, 113),
-    (115, 71, 90),
-    (99, 75, 71),
-    (174, 182, 189),
-]
-
-palette_match = {
-    "pink": ["lightblue", "darkblue", "grey", "white", "black"],
-    "red": ["lightblue", "darkblue", "grey", "white", "black"],
-    "orange": ["green", "lightblue", "darkblue", "whiite", "black"],
-    "beige": ["darkblue", "purple", "brown", "white", "black"],
-    "yellow": ["green", "darkblue", "white", "black"],
-    "green": ["orange", "purple", "white", "black"],
-    "lightblue": ["pink", "red", "orange", "white", "black"],
-    "darkblue": ["pink", "red", "orange", "white", "black"],
-    "purple": ["orange", "grey", "green", "white", "black"],
-    "brown": ["beige", "white", "black"],
-    "grey": ["pink", "red", "darkblue", "brown"]
-}
-
-palette_names = {
-    (255, 209, 223): "pink",
-    (210, 66, 71): "red",
-    (241, 89, 42): "orange",
-    (245, 239, 221): "beige",
-    (251, 214, 78): "yellow",
-    (0, 121, 71): "green",
-    (173, 216, 230): "lightblue",
-    (26, 72, 113): "darkblue",
-    (115, 71, 90): "purple",
-    (99, 75, 71): "brown",
-    (174, 182, 189): "grey"
-}
