--- ml_app/models/huggingface_cloth_segmentation/requirements.txt
@@ -1,7 +1,0 @@
-torch
-torchvision
-gradio
-gdown
-Pillow
-opencv-python
-numpy
