--- ml_app/requirements.txt
@@ -1,11 +1,0 @@
-torch
-flask
-pillow
-mysql-connector
-opencv-python-headless
-torchvision
-colorspacious
-colormath
-gdown
-gradio
-matplotlib
