+++ schema.sql
@@ -1,3 +1,5 @@
+CREATE DATABASE IF NOT EXISTS closetx;
+USE closetx;
 DROP TABLE IF EXISTS apparel;
 DROP TABLE IF EXISTS user;
 DROP USER IF EXISTS closetx;
